feat: client permission settings and permission seeder

+++ app/Http/Controllers/ClientController.php
@@ -9,6 +9,11 @@ use RealRashid\SweetAlert\Facades\Alert as Alert;
 
 class ClientController extends Controller
 {
+    public function __construct()
+    {
+       $this->middleware(['role:Vendedor|Administrador'])->except('index');
+    }
+
     /**
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
@@ -25,6 +30,7 @@ class ClientController extends Controller
     public function create(Request $request)
     {
         $client = new Client();
+        $this->authorize('create',$client);
         return view('client.create',compact('client'));
     }
 
@@ -34,7 +40,7 @@ class ClientController extends Controller
      */
     public function store(ClientStoreRequest $request)
     {
-
+        $this->authorize('create',new Client());
         $client = Client::create($request->validated());
         Alert::success('Cliente Guardado', '');
         return redirect()->route('clients.index');
@@ -47,10 +53,13 @@ class ClientController extends Controller
      */
     public function edit(Request $request, Client $client)
     {
+        $this->authorize('edit',$client);
         return view('client.create', compact('client'));
     }
 
     public function update(ClientStoreRequest $request, Client $client){
+
+        $this->authorize($client);
         $client->update($request->validated());
         Alert::success('Actualizado Correctamente');
         return redirect()->route('clients.index');
@@ -63,6 +72,7 @@ class ClientController extends Controller
      */
     public function destroy(Request $request, Client $client)
     {
+        $this->authorize($client);
         $client->delete();
 
         return redirect()->route('client.index');
